feat: storing the token in localstorag
