+++ src/app/admin/page.tsx
@@ -1,0 +1,8 @@
+const AdminPage = () => {
+  return (
+    <div>
+      <p>This is the admin page.</p>
+    </div>
+  );
+};
+export default AdminPage;
